Only do the "feature dns" DNS-lookup if/when necessary.
Spotted by:	Willy Tarreau <w@1wt.eu>
